+++ noodles-bam/CHANGELOG.md
@@ -4,8 +4,8 @@
 
 ### Changed
 
-  * bam/record: Vector fields are lazily decoded: read name, sequence, quality
-    scores, and data.
+  * bam/record: Vector fields are lazily decoded: cigar, read name, sequence,
+    quality scores, and data.
 
 ## 0.18.0 - 2022-04-14